feat: 存储空间重构、系统初始化与相关页面适配
- 存储空间模型与服务重构，移除扩容弹窗，更新管理页面

- 新增系统初始化接口与页面

- 用户/角色/权限/登录等相关 DTO/VO 与页面同步调整

- 更新测试与 Flyway 迁移

+++ app/src/main/java/com/fleyx/jcloud/model/vo/ChunkedUploadInitVo.java
@@ -22,7 +22,7 @@ public class ChunkedUploadInitVo implements Serializable {
     /**
      * 分片大小（字节）。
      */
-    private Long chunkSize;
+    private Integer chunkSize;
 
     /**
      * 总分片数。
